home moderador por la parte de reportes y ultimos reportes hecha

+++ src/main/java/com/iesfernandoaguilar/perezgonzalez/controller/Controller_Usuario.java
@@ -28,7 +28,7 @@ public class Controller_Usuario {
         this.controller.abrirPerfilUsuario(usuario);
     }
 
-    public void setCantReportes(int cantReportes){
+    public void setCantReportes(long cantReportes){
         this.Lbl_CantReportes.setText(String.valueOf(cantReportes));
     }
